Better Pop Up Blocker 2.1.4 August 11th, 2010 Minor bug fix release while I work on a "very interesting" Google Chrome extension.
More compatibility improvements for window.open.

Fixed a potential issue where a site may not be totally delisted from the whitelist the first time if there are multiple matching rules for the site's url.

Implemented more rigorous url matching.

Added a check so that Google Chrome 5.0 or later is required.